src/third_party/blink/web_tests/inspector-protocol/debugger/domdebugger-set-instrumentation-breakpoint-expected.txt

Tests the DOMDebugger.setInstrumentationBreakpoint API.

Running test: testScriptFirstStatement
{
    eventName : instrumentation:scriptFirstStatement
}

Running test: testRemoveScriptFirstStatement

Running test: testInvalidEventName
{
    id : <number>
    result : {
    }
    sessionId : <string>
}

Running test: testScriptFirstStatementAndPause
{
    reasons : [
        [0] : {
            reason : other
        }
        [1] : {
            auxData : {
                eventName : instrumentation:scriptFirstStatement
            }
            reason : EventListener
        }
    ]
}
